requirements.txt
numpy
gunicorn
scipy
sklearn
flask
pillow
